+++ Solution/MPEngine/Graphics/PostEffect/IntermediateRenderTarget.h
@@ -9,8 +9,8 @@ public:
 	IntermediateRenderTarget(class DeviceManager* device, class SwapChain* swapChain, class ResourceManager* rsManager);
 	~IntermediateRenderTarget() = default;
 
-	uint8_t PreProcess(ID3D12GraphicsCommandList* comList, uint8_t setHandleNumber, bool thisResource = true);
-	uint8_t PostProcess();
+	uint32_t PreProcess(ID3D12GraphicsCommandList* comList, uint32_t setHandleNumber, bool thisResource = true);
+	uint32_t PostProcess();
 
 private:
 	void CreatePipelineState() override;
